how-to-guides/data-versioning/scripts/requirements.txt
neptune-client
scikit-learn==0.24.1